+++ packages/app-builder-lib/src/node-module-collector/nodeModulesCollector.ts
@@ -7,6 +7,7 @@ import { exec, log } from "builder-util"
 export abstract class NodeModulesCollector {
   private nodeModules: NodeModuleInfo[]
   protected dependencyPathMap: Map<string, string>
+  protected allDependencies: Map<string, DependencyTree> = new Map()
 
   constructor(private readonly rootDir: string) {
     this.dependencyPathMap = new Map()
@@ -33,7 +34,7 @@ export abstract class NodeModulesCollector {
     return node
   }
 
-  private resolvePath(filePath: string) {
+  protected resolvePath(filePath: string) {
     try {
       const stats = fs.lstatSync(filePath)
       if (stats.isSymbolicLink()) {
@@ -47,7 +48,7 @@ export abstract class NodeModulesCollector {
     }
   }
 
-  public convertToDependencyGraph(tree: DependencyTree): DependencyGraph {
+  private convertToDependencyGraph(tree: DependencyTree): DependencyGraph {
     const result: DependencyGraph = { ".": {} }
 
     const flatten = (node: DependencyTree, parentKey = ".") => {
@@ -73,8 +74,19 @@ export abstract class NodeModulesCollector {
     return result
   }
 
+  getAllDependencies(tree: DependencyTree) {
+    const dependencies = tree.dependencies || {}
+    for (const [key, value] of Object.entries(dependencies)) {
+      if (value.dependencies && Object.keys(value.dependencies).length > 0) {
+        this.allDependencies.set(`${key}@${value.version}`, value)
+        this.getAllDependencies(value)
+      }
+    }
+  }
+
   abstract getCommand(): string
   abstract getArgs(): string[]
+  abstract removeNonProductionDependencie(tree: DependencyTree): void
 
   protected async getDependenciesTree(): Promise<DependencyTree> {
     const command = this.getCommand()
@@ -84,7 +96,18 @@ export abstract class NodeModulesCollector {
       shell: true,
     })
     const dependencyTree: DependencyTree | DependencyTree[] = JSON.parse(dependencies)
-    return Array.isArray(dependencyTree) ? dependencyTree[0] : dependencyTree
+
+    // pnpm returns an array of dependency trees
+    if (Array.isArray(dependencyTree)) {
+      const tree = dependencyTree[0]
+      if (tree.optionalDependencies) {
+        tree.dependencies = { ...tree.dependencies, ...tree.optionalDependencies }
+      }
+      return tree
+    }
+
+    // yarn and npm return a single dependency tree
+    return dependencyTree
   }
 
   private _getNodeModules(dependencies: Set<HoisterResult>, result: NodeModuleInfo[]) {
@@ -127,6 +150,8 @@ export abstract class NodeModulesCollector {
   public async getNodeModules(): Promise<NodeModuleInfo[]> {
     const tree = await this.getDependenciesTree()
     const realTree = this.getTreeFromWorkspaces(tree)
+    this.getAllDependencies(realTree)
+    this.removeNonProductionDependencie(realTree)
     const dependencyGraph = this.convertToDependencyGraph(realTree)
     const hoisterResult = hoist(this.transToHoisterTree(dependencyGraph), { check: true })
     this._getNodeModules(hoisterResult.dependencies, this.nodeModules)