+++ packages/app-builder-lib/src/node-module-collector/yarnNodeModulesCollector.ts
@@ -1,4 +1,5 @@
 import { NodeModulesCollector } from "./nodeModulesCollector"
+import { DependencyTree } from "./types"
 
 export class YarnNodeModulesCollector extends NodeModulesCollector {
   constructor(rootDir: string) {
@@ -10,6 +11,22 @@ export class YarnNodeModulesCollector extends NodeModulesCollector {
   }
 
   getArgs(): string[] {
-    return ["list", "--omit", "dev", "-a", "--json", "--long", "--silent"]
+    return ["list", "-a", "--include", "prod", "--include", "optional", "--omit", "dev", "--json", "--long", "--silent"]
+  }
+
+  removeNonProductionDependencie(tree: DependencyTree) {
+    const dependencies = tree.dependencies || {}
+    const _dependencies = tree._dependencies || {}
+    if (dependencies && Object.keys(dependencies).length === 0) {
+      tree.dependencies = this.allDependencies.get(`${tree.name}@${tree.version}`)?.dependencies || {}
+    }
+
+    for (const [key, value] of Object.entries(dependencies)) {
+      if (!_dependencies[key] || Object.keys(value).length === 0) {
+        delete dependencies[key]
+        continue
+      }
+      this.removeNonProductionDependencie(value)
+    }
   }
 }